requirements.txt
Django==3.0.8
django-solo==1.1.5